+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 This project is versioned based upon the version of [zigbee2mqtt](https://github.com/Koenkk/zigbee2mqtt). The versioning `X.Y.Z` of the stable release of this add-on will track that of zigbee2mqtt. If there are new releases without upgrades to the zigbee2mqtt version (i.e., changes to the add-on that occur between releases of zigbee2mqtt), an additional number will be added to indicate this (`X.Y.Z.A`, where `A` indicates a new versioned release).
 
+## [1.0.1](https://github.com/danielwelch/hassio-zigbee2mqtt/releases/tag/v1.1.0) - 2019-01-30
+### Changed
+- update zigbee2mqtt to version 1.1.0
+### Added
+- Includes new config options from zigbee2mqtt:
+    - `soft_reset_timeout`
+    - `last_seen`
+    - `elapsed`
+    - `network_key`
+    - `devices` (device-specific configuration)
+- Includes `overwrite` option, which can be set to `false` to prevent the add-on from overwriting corresponding fields on startup that are already set in the `configuration.yaml` file.
+
 ## [1.0.1](https://github.com/danielwelch/hassio-zigbee2mqtt/releases/tag/v1.0.1) - 2018-12-17
 ### Changed
 - update zigbee2mqtt to version 1.0.1